+++ src/QuickMarcEditor/QuickMarcEditorRows/FixedField/ContinuingResourceFixedField.js
@@ -19,6 +19,7 @@ const config = {
     {
       type: SUBFIELD_TYPES.STRING,
       name: 'Lang',
+      length: 3,
     },
     {
       type: SUBFIELD_TYPES.BYTE,
@@ -39,6 +40,7 @@ const config = {
     {
       type: SUBFIELD_TYPES.STRING,
       name: 'Ctry',
+      length: 3,
     },
     {
       name: 'S/L',
@@ -76,10 +78,12 @@ const config = {
     {
       name: 'Date1',
       type: SUBFIELD_TYPES.STRING,
+      length: 4,
     },
     {
       name: 'Date2',
       type: SUBFIELD_TYPES.STRING,
+      length: 4,
     },
   ],
 };