+++ src/QuickMarcEditor/QuickMarcEditorRows/FixedField/VisualMaterialFixedField.js
@@ -19,6 +19,7 @@ const config = {
     {
       name: 'Lang',
       type: SUBFIELD_TYPES.STRING,
+      length: 3,
     },
     {
       name: 'Form',
@@ -35,6 +36,7 @@ const config = {
     {
       name: 'Ctry',
       type: SUBFIELD_TYPES.STRING,
+      length: 3,
     },
     {
       type: SUBFIELD_TYPES.BYTE,
@@ -54,12 +56,14 @@ const config = {
       type: SUBFIELD_TYPES.BYTE,
     },
     {
-      type: SUBFIELD_TYPES.STRING,
       name: 'Date1',
+      type: SUBFIELD_TYPES.STRING,
+      length: 4,
     },
     {
       name: 'Date2',
       type: SUBFIELD_TYPES.STRING,
+      length: 4,
     },
   ],
 };