+++ src/QuickMarcEditor/utils.js
@@ -24,9 +24,12 @@ import {
   LEADER_DOCUMENTATION_LINKS,
   ELVL_BYTE,
   CREATE_BIB_RECORD_DEFAULT_FIELD_TAGS,
-  BIB_FIXED_FIELD_DEFAULT_VALUES,
+  BIB_FIXED_FIELD_DEFAULT_TYPE,
+  BIB_FIXED_FIELD_DEFAULT_BLVL,
+  DATE_ON_ENTERED_PLACEHOLDER,
 } from './constants';
 import { RECORD_STATUS_NEW } from './QuickMarcRecordInfo/constants';
+import { SUBFIELD_TYPES } from './QuickMarcEditorRows/BytesField';
 import getMaterialCharsFieldConfig from './QuickMarcEditorRows/MaterialCharsField/getMaterialCharsFieldConfig';
 import getPhysDescriptionFieldConfig from './QuickMarcEditorRows/PhysDescriptionField/getPhysDescriptionFieldConfig';
 import { FixedFieldFactory } from './QuickMarcEditorRows/FixedField';
@@ -205,10 +208,40 @@ const getCreateMarcRecordDefaultFields = (instanceRecord) => {
   });
 };
 
+export const fillEmptyFixedFieldValues = (marcType, type, blvl, field) => {
+  const fieldConfigByType = FixedFieldFactory
+    .getFixedFieldByType(
+      marcType,
+      type,
+      blvl,
+    )?.configFields ?? [];
+
+  return fieldConfigByType.reduce((fixedField, fieldConfig) => {
+    if (fixedField?.[fieldConfig.name]) {
+      return fixedField;
+    }
+
+    if (fieldConfig.type === SUBFIELD_TYPES.BYTE) {
+      return { ...fixedField, [fieldConfig.name]: '\\' };
+    } else if (fieldConfig.type === SUBFIELD_TYPES.BYTES) {
+      return { ...fixedField, [fieldConfig.name]: new Array(fieldConfig.bytes).fill('\\') };
+    } else if (fieldConfig.type === SUBFIELD_TYPES.STRING) {
+      return { ...fixedField, [fieldConfig.name]: new Array(fieldConfig.length).fill('\\').join('') };
+    }
+
+    return fixedField;
+  }, {
+    ...field?.content,
+    Type: type,
+    BLvl: blvl,
+    Entered: DATE_ON_ENTERED_PLACEHOLDER,
+  });
+};
+
 const getCreateBibMarcRecordDefaultFields = (instanceRecord) => {
   const contentMap = {
     '001': instanceRecord.hrid,
-    '008': BIB_FIXED_FIELD_DEFAULT_VALUES,
+    '008': fillEmptyFixedFieldValues(MARC_TYPES.BIB, BIB_FIXED_FIELD_DEFAULT_TYPE, BIB_FIXED_FIELD_DEFAULT_BLVL),
     '245': '$a ',
     '999': '',
   };
@@ -997,6 +1030,28 @@ export const autopopulateIndicators = (formValues) => {
   };
 };
 
+export const autopopulateFixedField = (formValues, marcType) => {
+  const { records } = formValues;
+
+  const leader = records.find(field => field.tag === LEADER_TAG);
+  const type = leader.content[6];
+  const blvl = leader.content[7];
+
+  return {
+    ...formValues,
+    records: records.map(field => {
+      if (!isFixedFieldRow(field)) {
+        return field;
+      }
+
+      return {
+        ...field,
+        content: fillEmptyFixedFieldValues(marcType, type, blvl, field),
+      };
+    }),
+  };
+};
+
 export const autopopulateSubfieldSection = (formValues, marcType = MARC_TYPES.BIB) => {
   const { records } = formValues;
 
@@ -1035,7 +1090,7 @@ export const autopopulateSubfieldSection = (formValues, marcType = MARC_TYPES.BI
   };
 };
 
-export const cleanBytesFields = (formValues, initialValues, marcType) => {
+export const cleanBytesFields = (formValues, marcType) => {
   const { records } = formValues;
 
   const cleanedRecords = records.map((field) => {
@@ -1055,7 +1110,7 @@ export const cleanBytesFields = (formValues, initialValues, marcType) => {
 
     if (isFixedFieldRow(field)) {
       fieldConfigByType = FixedFieldFactory
-        .getFixedFieldByType(marcType, field.content.Type, initialValues?.leader[7])?.configFields ?? [];
+        .getFixedFieldByType(marcType, field.content.Type, field.content.BLvl)?.configFields ?? [];
     }
 
     const content = Object.entries(field.content).reduce((acc, [key, value]) => {