+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * [UIQM-470](https://issues.folio.org/browse/UIQM-470) Add permission for editing MARC Holdings records.
 * [UIQM-419](https://issues.folio.org/browse/UIQM-419) Make Leader validation error message consistent for all MARC types.
 * [UIQM-465](https://issues.folio.org/browse/UIQM-465) Bib Rec. / Field 008 / Ctrl / Removed unused Ctrl field.
+* [UIQM-464](https://issues.folio.org/browse/UIQM-464) Create Orig Bib Record: Populate new record with 008.
 
 ## [6.0.2](https://github.com/folio-org/ui-quick-marc/tree/v6.0.2) (2023-03-30)